+++ src/Traits/Feed/HasCategoryTreeStructure.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace TimothyDC\LightspeedEcomProductFeed\Traits\Feed;
 
 use Illuminate\Support\Str;
+use TimothyDC\LightspeedEcomProductFeed\Traits\Feed\Base\HasCategoryInfo;
 
 /**
  * This trait will generate something like: