+++ tests/Feature/Feeds/SooqrFeedTest.php
@@ -264,18 +264,8 @@ class SooqrFeedTest extends TestCase
                     ],
                 ],
             ],
-            'specifications' => [
-                'node' => [
-                    [
-                        'title' => ['_cdata' => 'Packages'],
-                        'value' => ['_cdata' => 5],
-                    ],
-                    [
-                        'title' => ['_cdata' => 'Airtight'],
-                        'value' => ['_cdata' => 'Yes'],
-                    ],
-                ],
-            ],
+            'packages' => [['_cdata' => 5]],
+            'airtight' => [['_cdata' => 'Yes']],
         ];
     }
 
@@ -321,18 +311,8 @@ class SooqrFeedTest extends TestCase
         </values>
       </node>
     </filters>
-    <specifications>
-      <node>
-        <title><![CDATA[Packages]]></title>
-        <value>
-          <_cdata>5</_cdata>
-        </value>
-      </node>
-      <node>
-        <title><![CDATA[Airtight]]></title>
-        <value><![CDATA[Yes]]></value>
-      </node>
-    </specifications>
+    <packages><![CDATA[5]]></packages>
+    <airtight><![CDATA[Yes]]></airtight>
   </product>
 </products>
 ';