+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to generate a flat representation of the category tree structure via `HasCategoryTreeStructureFlat`-trait
 - Option to generate a flat representation of the specification data via `HasSpecificationsAsNodes`-trait
 - Option to skip certain data while processing categories, filters, images, specifications or variants
+- Custom `Str::xmlNode` function to convert a string to an XML node safe name
 
 ## [v1.7.2] - 2020-08-03
 ### Changed