+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.8.0] - 2020-08-11
 ### Added
 - Option to generate a flat representation of the category tree structure via `HasCategoryTreeStructureFlat`-trait
 - Option to generate a flat representation of the specification data via `HasSpecificationsAsNodes`-trait
@@ -115,7 +117,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Commands to create, list and remove product feed via `artisan`-commands
 - Option to overwrite XML data structure via interface binding
 
-[Unreleased]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.7.2...HEAD
+[Unreleased]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.8.0...HEAD
+[v1.8.0]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.7.2...v1.8.0
 [v1.7.2]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.7.1...v1.7.2
 [v1.7.1]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.7.0...v1.7.1
 [v1.7.0]: https://github.com/timothydc/laravel-lightspeed-ecom-product-feeds/compare/v1.6.0...v1.7.0